TestModels/aws-sdks/kms/runtimes/python/pyproject.toml
[tool.poetry]
name = "amazon-cryptography-internal-dafny-shim-kms"
version = "0.1.0"
description = ""
authors = ["AWS <aws-arg-dafny@amazon.com>"]
packages = [
    { include = "com_amazonaws_kms", from = "src" }
]
# Include all of the following .gitignored files in package distributions,
# even though it is not included in version control
include = ["**/smithygenerated/**/*.py", "**/dafnygenerated/*.py"]

[tool.poetry.dependencies]
python = "^3.11.0"
boto3 = "^1.28.38"
# TODO: Depend on PyPi once Smithy-Python publishes their Python package there
smithy-python = { path = "../../../../../submodules/smithy-python/python-packages/smithy-python", develop = false}
standard-library = {path = "../../../../dafny-dependencies/StandardLibrary/runtimes/python", develop = false}
DafnyRuntimePython = "^4.4.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
